requirements.txt
Pillow==11.2.1
django==4.2.20
djangorestframework==3.16.0
django-filter==25.1